Creates validation on requests

+++ bootcamp-api/routes/tasks.js
@@ -1,3 +1,6 @@
+const { body, param, validationResult } = require("express-validator/check");
+const { matchedData } = require("express-validator/filter");
+
 module.exports = app => {
   const Tasks = app.db.models.Tasks;
 
@@ -14,20 +17,40 @@ module.exports = app => {
       }
     })
 
-    .post(async (req, res) => {
-      try {
-        const task = await Tasks.create(red.body);
-        res.json(task);
-      } catch (error) {
-        console.log(error);
-        res.status(500).json({ msg: "Unexpected error" });
+    .post(
+      [
+        body("title", "Required field").exists(),
+        body("title", "Invalid length")
+          .trim()
+          .isLength({ min: 1, max: 255 })
+      ],
+      async (req, res) => {
+        try {
+          const errors = validationResult(req);
+
+          if (!errors.isEmpty()) {
+            return res.status(400).json({ errors: errors.array() });
+          }
+
+          const task = await Tasks.create(matchedData(req));
+          res.json(task);
+        } catch (error) {
+          console.log(error);
+          res.status(500).json({ msg: "Unexpected error" });
+        }
       }
-    });
+    );
 
   app
     .route("/tasks")
-    .get(async (req, res) => {
+    .get([param("id", "Not an integer").isInt()], async (req, res) => {
       try {
+        const errors = validationResult(req);
+
+        if (!errors.isEmpty()) {
+          return res.status(400).json({ errors: errors.array() });
+        }
+
         const task = await Tasks.findById(req.params.id);
 
         if (task) {
@@ -41,23 +64,47 @@ module.exports = app => {
       }
     })
 
-    .put(async (req, res) => {
-      try {
-        await Tasks.update(req.body, {
-          where: {
-            id: req.params.id
+    .put(
+      [
+        param("id", "Not an integer").isInt(),
+        body("title", "Required field").exists(),
+        body("title", "Invalid length"),
+        body("done", "Required field").exists(),
+        body("done", "Not a boolean").isBoolean()
+      ],
+      async (req, res) => {
+        try {
+          const errors = validationResult(req);
+
+          if (!errors.isEmpty()) {
+            return res.status(400).json({ errors: errors.array() });
           }
-        });
 
-        res.sendStatus(204);
-      } catch (error) {
-        console.log(error);
-        res.status(500).json({ msg: "Unexpected error" });
+          await Tasks.update(matchedData(req), {
+            where: {
+              id: req.params.id
+            }
+          });
+
+          res.sendStatus(204);
+        } catch (error) {
+          console.log(error);
+          res.status(500).json({ msg: "Unexpected error" });
+        }
       }
-    })
+    )
 
-    .delete(async (req, res) => {
+    .delete(
+      [
+        param("id", "Not an integer").isInt()
+      ], async (req, res) => {
       try {
+        const errors = validationResult(req);
+
+        if (!errors.isEmpty()) {
+          return res.status(400).json({ errors: errors.array() });
+        }
+
         await Tasks.destroy({
           where: {
             id: req.params.id
